+++ lib/inquisition.rb
@@ -16,7 +16,8 @@ require 'inquisition/brakeman/runner'
 require 'inquisition/bundler/audit/runner'
 require 'inquisition/fasterer/runner'
 require 'inquisition/rails_best_practices/runner'
-require 'inquisition/rails_best_practices/vulnerability'
+require 'inquisition/rails_best_practices/issue'
+require 'inquisition/rails_best_practices/categorizer'
 require 'inquisition/rubocop'
 require 'inquisition/lol_dba/runner'
 require 'inquisition/fasterer/runner'
